refactored data structures for singel source of truth for front/backend
