+++ .eslintrc.json
@@ -10,7 +10,7 @@
           "error",
           {
             "enforceBuildableLibDependency": true,
-            "allow": [],
+            "allow": ["../../../shared/**"],
             "depConstraints": [
               {
                 "sourceTag": "*",
