+++ apps/backend/src/applications/applications.controller.spec.ts
@@ -9,8 +9,8 @@ import {
   Semester,
   StageProgress,
   ReviewStatus,
-} from './types';
-import { UserStatus } from '../users/types';
+} from '../../../shared/types/application.types';
+import { UserStatus } from '../../../shared/types/user.types';
 import { userFactory } from '../testing/factories/user.factory';
 import { AuthService } from '../auth/auth.service';
 import { UsersService } from '../users/users.service';
