--- apps/backend/src/auth/dtos/sign-in.response.dto.ts
@@ -1,5 +1,0 @@
-export class SignInResponseDto {
-  accessToken: string;
-
-  refreshToken: string;
-}
