+++ apps/backend/src/reviews/review.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
 import { Application } from '../applications/application.entity';
-import { ApplicationStage } from '../applications/types';
+import { ApplicationStage } from '../../../shared/types/application.types';
 
 @Entity()
 export class Review {
