--- apps/backend/src/users/types.ts
@@ -1,34 +1,0 @@
-// TODO Probably want these types to be available to both the frontend and backend in a "common" folder
-
-export enum UserStatus {
-  MEMBER = 'Member',
-  RECRUITER = 'Recruiter',
-  ADMIN = 'Admin',
-  ALUMNI = 'Alumni',
-  APPLICANT = 'Applicant',
-}
-
-export enum Team {
-  SFTT = 'Speak for the Trees',
-  CONSTELLATION = 'Constellation',
-  JPAL = 'J-PAL',
-  BREAKTIME = 'Breaktime',
-  GI = 'Green Infrastructure',
-  CI = 'Core Infrastructure',
-  EBOARD = 'E-Board',
-}
-
-export enum Role {
-  DIRECTOR_OF_ENGINEERING = 'Director of Engineering',
-  DIRECTOR_OF_PRODUCT = 'Director of Product',
-  DIRECTOR_OF_FINANCE = 'Director of Finance',
-  DIRECTOR_OF_MARKETING = 'Director of Marketing',
-  DIRECTOR_OF_RECRUITMENT = 'Director of Recruitment',
-  DIRECTOR_OF_OPERATIONS = 'Director of Operations',
-  DIRECTOR_OF_EVENTS = 'Director of Events',
-  DIRECTOR_OF_DESIGN = 'Director of Design',
-  PRODUCT_MANAGER = 'Product Manager',
-  PRODUCT_DESIGNER = 'Product Designer',
-  TECH_LEAD = 'Technical Lead',
-  DEVELOPER = 'Developer',
-}
