+++ apps/backend/src/users/user.entity.ts
@@ -11,7 +11,7 @@ import {
 } from 'class-validator';
 import { Entity, Column, OneToMany } from 'typeorm';
 import { Application } from '../applications/application.entity';
-import { Role, Team, UserStatus } from './types';
+import { Role, Team, UserStatus } from '../../../shared/types/user.types';
 
 @Entity()
 export class User {
