+++ apps/shared/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Main export file for shared types and DTOs
+ */
+
+// Export all type files
+export * from './types/application.types';
+export * from './types/user.types';
+export * from './types/auth.types';
+export * from './types/review.types';
+
+// Export all DTO files
+export * from './dto/application.dto';
+export * from './dto/user.dto';
+export * from './dto/auth.dto';
