+++ tsconfig.base.json
@@ -14,7 +14,9 @@
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
-    "paths": {}
+    "paths": {
+      "@shared/*": ["apps/shared/*"]
+    }
   },
   "exclude": ["node_modules", "tmp"]
 }
